+++ tileserver/maptiles/globals.go
@@ -2,5 +2,5 @@ package maptiles
 
 const (
 	SERVER_NAME string = "MapnikServer"
-    VERSION string = "0.1.5"
-5
+	VERSION     string = "0.1.5"
+)